[BOOK-96] refactor: apis,domain - 패키지구조 변경 (#29)
