--- apis/src/main/kotlin/org/yapp/apis/util/AuthUtils.kt
@@ -1,13 +1,0 @@
-package org.yapp.apis.util
-
-import java.util.*
-
-object AuthUtils {
-    fun extractUserIdFromAuthHeader(
-        authorizationHeader: String,
-        getUserId: (String) -> UUID
-    ): UUID {
-        val token = authorizationHeader.removePrefix("Bearer ").trim()
-        return getUserId(token)
-    }
-}
