+++ domain/src/main/kotlin/org/yapp/domain/user/ProviderType.kt
@@ -1,4 +1,4 @@
-package org.yapp.domain.auth
+package org.yapp.domain.user
 
 /**
  * Enum representing different authentication providers.
