+++ infra/src/main/kotlin/org/yapp/infra/user/entity/UserEntity.kt
@@ -3,7 +3,7 @@ package org.yapp.infra.user.entity
 import jakarta.persistence.*
 import org.hibernate.annotations.JdbcTypeCode
 import org.hibernate.annotations.SQLDelete
-import org.yapp.domain.auth.ProviderType
+import org.yapp.domain.user.ProviderType
 import org.yapp.domain.common.BaseTimeEntity
 import org.yapp.domain.user.User
 import java.sql.Types
